FuelUpdated

+++ GoldRate/Models/mGoldRate.cs
@@ -17,13 +17,13 @@ namespace GoldRate.Models
         [Column(name:"cGoldCarat", TypeName ="Varchar(20)",Order =1)]
         public string GoldCarat { get; set; }
         [Column(name: "rGoldPriceMorning", TypeName = "decimal(9,3)", Order = 2)]
-        public decimal GoldPriceMorning { get; set; }
+        public double GoldPriceMorning { get; set; }
         [Column(name: "rGoldPriceAfternoon", TypeName = "decimal(9,3)", Order = 2)]
-        public decimal GoldPriceAfternoon { get; set; }
+        public double GoldPriceAfternoon { get; set; }
         [Column(name: "rGoldPriceEvening", TypeName = "decimal(9,3)", Order = 2)]
-        public decimal GoldPriceEvening { get; set; }
+        public double GoldPriceEvening { get; set; }
         [Column(name: "rGoldPriceYesterday", TypeName = "decimal(9,3)", Order = 2)]
-        public decimal GoldPriceYesterday { get; set; }
+        public double GoldPriceYesterday { get; set; }
         [Column(name: "dUpdatedDateTime", TypeName = "datetime", Order = 3)]
         public DateTime? UpdatedDateTime { get; set; }
     }
